Create BIBtoRDA_DumbFix
'MACRO <DumbFix_toRDA> (actions are listed not in the order they are performed):
'Login if not online
' FIXED FIELDS:
'Fixed Fields - make sure that they are displayed at Top or Bottom
'Fixed Fields (ELvl, Srce, Desc) - set them according to RDA or leave as it is if the record is PCC\LCODE
'Fixed Field <Ills> - set for Illustrations if present in 300
'Fixed Field <Ills> - set for maps if present in 300
'Fixed Field <Indx> - set for index if present in 504
'Fixed Field <Cont> - set for bibliograhpy if present in 504
'Fixed Field <LitF> - if a keyword, associated with a LITERARY WORK, is found in the record,  pop-up menu appears with possible choices
'Fixed Field <Biog> - if a keyword, associated with a BIOGRAPHY, found in the record,  pop-up menu appears with possible choices
' FIELDS:
'040 - check and insert if missing\wrong (Stanford) Language and Cataloging codes ($b eng $e rda)
'043 - check for errors and add if missing, based on LCSH contents
'050 - check if missing, fix formating errrors, sincronize the date with 264 and Fixed Field <Dates>
'100 - add <, $e author> if missing and control
'260 - change all to 264/1
'300 - add\remove <.> after <cm> based on 490 presence
'336, 337, 338 - check for errors and add if missing, based on the type of cataloged material (book, online, serial)
'830 - add if missing and 490 is present (with a paired field and validation).
'100, 245, 260, 264, 490, 600, 700, 830 - Fix single character Diacritics (in Stanford records - STF)
' REPLACE:
'  <double space> with <single space>
'  < p. > with < pages >
'  <(p. > with <(pages >
'  < pp. > with < pages >
'  <(pp. > with <(pages >
'  <ill.> with <illustrations>
'  Validate record
'TO BE ADDED:
'Fixed Field <Cont> -  if a keyword, associated with a DICTIONARY, is found in the record, set for dictionary
'Fixed Field <Cont> -  if a keyword, associated with a CATALOG, is found in the record, set for dictionary
'Hijy to Gregorian dates converter